The R_forPCs file worked great on my mac!!
line 9-10; can you check the objectives/activity numbers/letters?

line 58: I didn’t edit this in the script (the sim_folder/setwd command) but just did this manually in R studio- a reminder to you that there is some extra annotation around this line that should probably be edited

line 87: look up where your lake is (I think that they did this earlier to compare land uses, but I like leaving it in!)

line 136: my addition- could be deleted or not, but I see it as a way to slow them down and look through output vs blindly running code!

line 154: do we need to tell them to work with the other team of partners working on another lake in addition to their partner? Otherwise it seems a bit strange on line 242 for them to be working with their partner AND another pair.

lines 238-250: do you want them to plot the baseline & climate line plots here to answer question c? this also sets them up for the Activity  C factorial comparison.

line 260: same as 154

409: I wrote this up- please edit!